extract config & authn services
