+++ packages/unraid-api-plugin-connect/src/connection-status/cloud.dto.ts
@@ -1,6 +1,6 @@
 import { Field, Int, ObjectType } from '@nestjs/graphql';
 
-import { MinigraphStatus } from '../model/my-servers-config.model.js';
+import { MinigraphStatus } from '../config/my-servers.config.js';
 
 @ObjectType()
 export class ApiKeyResponse {
