+++ packages/unraid-api-plugin-connect/src/event-handler/wan-access.events.ts
@@ -2,8 +2,8 @@ import { Injectable, Logger } from '@nestjs/common';
 import { ConfigService } from '@nestjs/config';
 import { OnEvent } from '@nestjs/event-emitter';
 
+import { ConfigType } from '../config/connect.config.js';
 import { EVENTS } from '../helper/nest-tokens.js';
-import { ConfigType } from '../model/connect-config.model.js';
 import { NetworkService } from '../service/network.service.js';
 
 @Injectable()
