+++ packages/unraid-api-plugin-connect/src/job/timeout-checker.job.ts
@@ -3,8 +3,8 @@ import { SchedulerRegistry } from '@nestjs/schedule';
 
 import { isDefined } from 'class-validator';
 
+import { MinigraphStatus } from '../config/connect.config.js';
 import { ONE_MINUTE_MS, THREE_MINUTES_MS } from '../helper/generic-consts.js';
-import { MinigraphStatus } from '../model/connect-config.model.js';
 import { MothershipConnectionService } from '../service/connection.service.js';
 import { DynamicRemoteAccessService } from '../service/dynamic-remote-access.service.js';
 import { MothershipSubscriptionHandler } from '../service/mothership-subscription.handler.js';
