+++ packages/unraid-api-plugin-connect/src/resolver/connect.resolver.ts
@@ -9,7 +9,7 @@ import {
     UsePermissions,
 } from '@unraid/shared/use-permissions.directive.js';
 
-import { ConfigType, ConnectConfig, DynamicRemoteAccessType } from '../model/connect-config.model.js';
+import { ConfigType, ConnectConfig, DynamicRemoteAccessType } from '../config/connect.config.js';
 import { Connect, ConnectSettings, DynamicRemoteAccessStatus } from '../model/connect.model.js';
 
 @Resolver(() => Connect)
