+++ packages/unraid-api-plugin-connect/src/service/connect-settings.service.ts
@@ -21,10 +21,10 @@ import type {
     RemoteAccess,
     SetupRemoteAccessInput,
 } from '../model/connect.model.js';
+import { ConnectApiKeyService } from '../authn/connect-api-key.service.js';
+import { ConfigType, MyServersConfig } from '../config/connect.config.js';
 import { EVENTS } from '../helper/nest-tokens.js';
-import { ConfigType, MyServersConfig } from '../model/connect-config.model.js';
 import { DynamicRemoteAccessType, WAN_ACCESS_TYPE, WAN_FORWARD_TYPE } from '../model/connect.model.js';
-import { ConnectApiKeyService } from './connect-api-key.service.js';
 import { DynamicRemoteAccessService } from './dynamic-remote-access.service.js';
 import { NetworkService } from './network.service.js';
