+++ packages/unraid-api-plugin-connect/src/service/connection.service.ts
@@ -6,8 +6,8 @@ import type { OutgoingHttpHeaders } from 'node:http2';
 import { Subscription } from 'rxjs';
 import { bufferTime, filter } from 'rxjs/operators';
 
+import { ConnectionMetadata, MinigraphStatus, MyServersConfig } from '../config/connect.config.js';
 import { EVENTS } from '../helper/nest-tokens.js';
-import { ConnectionMetadata, MinigraphStatus, MyServersConfig } from '../model/connect-config.model.js';
 
 interface MothershipWebsocketHeaders extends OutgoingHttpHeaders {
     'x-api-key': string;
