+++ packages/unraid-api-plugin-connect/src/service/dynamic-remote-access.service.ts
@@ -3,14 +3,14 @@ import { ConfigService } from '@nestjs/config';
 
 import { URL_TYPE } from '@unraid/shared/network.model.js';
 
-import { ONE_MINUTE_MS } from '../helper/generic-consts.js';
 import {
     AccessUrlObject,
     ConfigType,
     DynamicRemoteAccessState,
     DynamicRemoteAccessType,
     makeDisabledDynamicRemoteAccessState,
-} from '../model/connect-config.model.js';
+} from '../config/connect.config.js';
+import { ONE_MINUTE_MS } from '../helper/generic-consts.js';
 import { StaticRemoteAccessService } from './static-remote-access.service.js';
 import { UpnpRemoteAccessService } from './upnp-remote-access.service.js';
